requirements.txt
colorama==0.4.6
PyJWT==2.6.0
requests==2.28.1
